클래스 분리 - 수강생, 점수 관리 분리
